Ajout du token dans les requêtes

+++ api/routes/parties_routes.js
@@ -1,13 +1,22 @@
 const express = require('express');
 const router = express.Router();
+const jwt = require('jsonwebtoken');
+const GameCode = require('../utils/code_generator');
+const {Base64} = require("js-base64");
 
 const Partie = require('../models/Partie');
 const User = require('../models/User');
-const Base64 = require('../utils/base64_encoder');
+
 
 //GET -> obtenir les parties d'un utilisateur / Obtenir tous les parties
 router.get('/', async (req,res) => {
+    const auth = req.header('Authorization');
+    if(auth === undefined || auth === null || auth === '') return res.status(400).json({message:'Token not found'});
+    const hash = auth.split(' ')[1];
+    if(hash === undefined || hash === null || hash === '') return res.status(400).json({message:'Token not found'});
+
     try {
+        const verify = jwt.verify(auth.split(' ')[1], Base64.encode('racingmazeapitoken'));
         let foundParties;
         if (req.query.userId != null) {
             foundParties = await Partie.find({'users.user._id': req.query.userId});
@@ -23,7 +32,13 @@ router.get('/', async (req,res) => {
 
 //GET -> Obtenir les parties publiques
 router.get('/publiques', async (req,res) => {
+    const auth = req.header('Authorization');
+    if(auth === undefined || auth === null || auth === '') return res.status(400).json({message:'Token not found'});
+    const hash = auth.split(' ')[1];
+    if(hash === undefined || hash === null || hash === '') return res.status(400).json({message:'Token not found'});
+
     try {
+        const verify = jwt.verify(auth.split(' ')[1], Base64.encode('racingmazeapitoken'));
         let foundParties = await Partie.find({type: 1});
         res.json(foundParties);
     }catch (error) {
@@ -34,7 +49,13 @@ router.get('/publiques', async (req,res) => {
 
 //GET -> obtenir une partie par id
 router.get('/:id/id', async (req,res) => {
+    const auth = req.header('Authorization');
+    if(auth === undefined || auth === null || auth === '') return res.status(400).json({message:'Token not found'});
+    const hash = auth.split(' ')[1];
+    if(hash === undefined || hash === null || hash === '') return res.status(400).json({message:'Token not found'});
+
     try {
+        const verify = jwt.verify(auth.split(' ')[1], Base64.encode('racingmazeapitoken'));
         let foundPartie = await Partie.findById(req.params.id);
         res.json(foundPartie);
     } catch (error) {
@@ -44,7 +65,13 @@ router.get('/:id/id', async (req,res) => {
 
 //GET -> obtenir une partie par code
 router.get('/:id/code', async (req,res) => {
+    const auth = req.header('Authorization');
+    if(auth === undefined || auth === null || auth === '') return res.status(400).json({message:'Token not found'});
+    const hash = auth.split(' ')[1];
+    if(hash === undefined || hash === null || hash === '') return res.status(400).json({message:'Token not found'});
+
     try {
+        const verify = jwt.verify(auth.split(' ')[1], Base64.encode('racingmazeapitoken'));
         let foundPartie = await Partie.findOne({code:req.params.id});
         res.json(foundPartie);
     } catch (error) {
@@ -54,30 +81,37 @@ router.get('/:id/code', async (req,res) => {
 
 //POST -> créer une partie
 router.post('/',async (req, res) => {
-    let foundUser = await User.findById(req.body.user.id);
-    let partie = new Partie({
-        nom: req.body.nom,
-        difficulte: req.body.difficulte,
-        type: req.body.type,
+    const auth = req.header('Authorization');
+    if(auth === undefined || auth === null || auth === '') return res.status(400).json({message:'Token not found'});
+    const hash = auth.split(' ')[1];
+    if(hash === undefined || hash === null || hash === '') return res.status(400).json({message:'Token not found'});
 
-        users: {
-            user:{
-                _id: foundUser._id,
-                prenom: foundUser.prenom,
-                nom: foundUser.nom,
-                username: foundUser.username,
-                personnage: foundUser.personnage
-            },
-            type: 1,
-            place: null,
-            score: null,
-            temps: null
-        }
-    });
-    if(req.body.type === 2){
-        partie.code = Base64.encoder(partie._id);
-    }
     try{
+        const verify = jwt.verify(auth.split(' ')[1], Base64.encode('racingmazeapitoken'));
+        let foundUser = await User.findById(verify.id);
+        let partie = new Partie({
+            nom: req.body.nom,
+            difficulte: req.body.difficulte,
+            type: req.body.type,
+            code: null,
+            users: {
+                user:{
+                    _id: foundUser._id,
+                    prenom: foundUser.prenom,
+                    nom: foundUser.nom,
+                    username: foundUser.username,
+                    personnage: foundUser.personnage
+                },
+                type: 1,
+                place: null,
+                score: null,
+                temps: null
+            }
+        });
+        if(req.body.type === 2){
+            partie.code = GameCode.generate(partie._id);
+        }
+
         let savedPartie = await partie.save();
         res.json(savedPartie);
     } catch(error){
@@ -87,7 +121,16 @@ router.post('/',async (req, res) => {
 
 //DELETE -> supprimer une partie
 router.delete('/:id', async (req, res) => {
+    const auth = req.header('Authorization');
+    if(auth === undefined || auth === null || auth === '') return res.status(400).json({message:'Token not found'});
+    const hash = auth.split(' ')[1];
+    if(hash === undefined || hash === null || hash === '') return res.status(400).json({message:'Token not found'});
+
     try {
+        const verify = jwt.verify(auth.split(' ')[1], Base64.encode('racingmazeapitoken'));
+        let actualUser = await User.findById(verify.id);
+        let actualPartie = await Partie.findById(req.params.id);
+
         let deletedPartie = await Partie.deleteOne({_id: req.params.id});
         res.json(deletedPartie);
     } catch (error) {
@@ -97,14 +140,21 @@ router.delete('/:id', async (req, res) => {
 
 //PUT -> modifier les données d'une partie
 router.put('/:id', async (req, res) => {
+    const auth = req.header('Authorization');
+    if(auth === undefined || auth === null || auth === '') return res.status(400).json({message:'Token not found'});
+    const hash = auth.split(' ')[1];
+    if(hash === undefined || hash === null || hash === '') return res.status(400).json({message:'Token not found'});
+
     try{
+        const verify = jwt.verify(auth.split(' ')[1], Base64.encode('racingmazeapitoken'));
         let modifiedPartie = await Partie.updateOne(
             {_id: req.params.id},
             {
                 $set: {
                     nom: req.body.nom,
                     difficulte: req.body.difficulte,
-                    type: req.body.type
+                    type: req.body.type,
+                    code: (req.body.type === 2) ? GameCode.generate(req.params.id) : null,
                 }
             }
         );
@@ -116,28 +166,40 @@ router.put('/:id', async (req, res) => {
 
 //PUT -> ajouter un utilisateur à une partie
 router.put('/:id/user', async (req, res) => {
+    const auth = req.header('Authorization');
+    if(auth === undefined || auth === null || auth === '') return res.status(400).json({message:'Token not found'});
+    const hash = auth.split(' ')[1];
+    if(hash === undefined || hash === null || hash === '') return res.status(400).json({message:'Token not found'});
+
     try{
-        let foundUser = await User.findById(req.body.user.id);
-        let modifiedPartie = await Partie.updateOne(
-            {_id: req.params.id},
-            {
-                $push:{
-                    users: {
-                        user:{
-                            _id: foundUser._id,
-                            prenom: foundUser.prenom,
-                            nom: foundUser.nom,
-                            username: foundUser.username,
-                            personnage: foundUser.personnage
-                        },
-                        place: null,
-                        score: null,
-                        temps: null
+        const verify = jwt.verify(auth.split(' ')[1], Base64.encode('racingmazeapitoken'));
+        let foundUser = await User.findById(verify.id);
+        let partie = await Partie.findOne({$and:[{"_id":req.params.id},{ "users.user._id" : foundUser._id}]});
+        console.log(partie);
+        if(partie === null){
+            let modifiedPartie = await Partie.updateOne(
+                {_id: req.params.id},
+                {
+                    $push:{
+                        users: {
+                            user:{
+                                _id: foundUser._id,
+                                prenom: foundUser.prenom,
+                                nom: foundUser.nom,
+                                username: foundUser.username,
+                                personnage: foundUser.personnage
+                            },
+                            place: null,
+                            score: null,
+                            temps: null
+                        }
                     }
                 }
-            }
-        );
-        res.json(modifiedPartie);
+            );
+            res.json(modifiedPartie);
+        } else {
+            res.status(400).json({message:'The user is already in the game'});
+        }
     } catch (error) {
         res.status(400).json({message:error});
     }
@@ -160,14 +222,5 @@ router.put('/:id/labyrinthe', async (req,res) => {
     }
 });
 
-//POST -> Se connecter
-router.post('/login',async (req,res) => {
-
-});
-
-//DELETE -> Se déconnecter
-router.post('/logout',async (req,res) => {
-
-});
 
 module.exports = router;
